fixed entity spawn

+++ src/network/protocol/packet/AddEntityPacket.php
@@ -22,7 +22,7 @@ class AddEntityPacket extends RakNetDataPacket{
 	public function encode(){
 		$this->reset();
 		$this->putInt($this->eid);
-		$this->putByte($this->type);
+		$this->putInt($this->type);
 		$this->putFloat($this->x);
 		$this->putFloat($this->y);
 		$this->putFloat($this->z);
